add support for external drivers to be installed
this commit allows an external driver to be installed to generate files
using xivo-confgend.

An example can be found here

https://gist.github.com/pcadottemichaud/e684b46234254dea3f33c0015cf48f47

To choose the driver to use modify the configuration to select the
desired plugin

in /etc/xivo-confgend/conf.d/my_sip_conf.yml

```
plugins:
  asterisk.sip.conf: dummy
```

this declaration will find any installed plugins with the entry_point
"confgend.asterisk.sip.conf"

if no plugin is found for the specified "file name" the old system is
used.