return cached content only on error
It's valid for some handlers to return an empty string (e.g. for
asterisk/queuerules.conf), and in that case xivo-confgend SHOULD NOT
return the previously cached content, but instead return an empty
string. Returning the cached content is wrong in this scenario.

Also, since an handler is not expected to return the value None, only
return the cached content when an exception is raised by the handler.
This is more in line with the previous (xivo 16.09) behavior, where the
cached content was returned either if an exception was raised or the
handler returned None (but no handler were actually returning None, so
this feature was unused).