sip.conf: add Twilio config generation helper
Twilio's trunk doesn't match completly the trunk notion of XiVO, and
this option helps lessen the burden of configuration on the user.

That said, there is some limitations regarding Twilio trunks and XiVO:
for example, if you have multiple entities on your XiVO and you would
like to use a Twilio trunk in each entity, enabling the option on the
XiVO trunks won't be enough for incoming calls to make sense, i.e. all
incoming calls from Twilio will enter a single context (in a single
entity). You'll need extra configuration at the XiVO global level to
make it work.