pjsip: allow all global options to be specified in the sip config
The SIP General Settings that match a global option will be added to the global
pjsip section.

The endpoint_identifier_order is an option that was manually mapped with a
default value. I remove the manual addition of that field and used setdefault
instead of assigning to an new variable.